requirements.txt
fastapi==0.114.2
uvicorn==0.30.6
sqlalchemy==1.4.51
psycopg2-binary==2.9.9
pandas==2.2.2
python-dotenv==1.0.1
pydantic==2.9.1
python-multipart==0.0.9